chore: rename package name
