+++ Makefile
@@ -1,5 +1,5 @@
 APP_VERSION := $(shell grep -oP '(?<=^version = ")[^"]*' pyproject.toml)
-APP_DIR := aiostatsd
+APP_DIR := asyncstatsd
 NPROCS = $(shell grep -c 'processor' /proc/cpuinfo)
 MAKEFLAGS += -j$(NPROCS)
 PYTEST_FLAGS := --failed-first -x
